pyproject.toml
[tool.poetry]
name = "fun-api"
version = "0.1.0"
description = ""
authors = ["SMT <smichaelthomas@wealthsimple.com>"]
readme = "README.md"
packages = [{include = "fun_api"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.4"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
